Added DirectoryWatcher
The DirectoryWatcher has a list of commands (basically callbacks) that are executed depending on the type of event in the swtich, FileCreatedEvent, FileDeletedEvent and so on, i won't implement directory cause it would be redundant for now... maybe later i'll find some edge case that will force me to do that.

I get the filename and path to the files from the watcher, more things i can do with the asset manager, like hot reloading scripts or textures and so on.
Currently it looks like the directory watcher fires extra events and must make sure i get things right with hot reloading.

Directory navigation in the Assets window, will implement drag and drop to add elements to the current scene