~ Instanced rendering ~
Managed to implement instanced rendering, uploading the geometry once per model and then if we want to draw the same geometry but 100 times, we will use a single draw call to draw them all.
There is a catch with glDrawElementsInstanced, it will need a transformation matrix, one per instance so the geometry could be the same but the instance matrix should be different.

This works verry well for things like grass, or trees where we have thousands of them and need to render them efficiently.

- vertex shader is update to support instance matrix
- Model class has new functions tied to those instances, for sure the instance matrices vector SHOULD NOT live in the model class, I think at first glance that we need some counter to determine how many of the same object we must draw and then setup the instancing data

800 thousand cubes with a single texture ~100fps