Man this git is shit ( did an opsie with the git) -_- (#3)
* Tooltip hover for scene hierarchy entities

Draws the texture, path, name and size of the texture component, i can specialize this popup.
I can pass in another draw function the entity and i will try_get<TComponent> there and depending on that i draw the texture info or model info

* Added EntityPropertiesWindow and frame time display

- Added ChangeEntityEvent class to handle entity selection changes from SceneHierarchyWindow
  - Currently dispatched to EntityPropertiesWindow as the only listener
- Added frame time display to the PerformanceWindow (GPU provider coming later)
- Renamed Event class to IEvent
- Removed .getInstance() from SceneManager (static inline methods are used now)

* Event Manager -> Event Dispatcher ( new implementation using EnTT)

- removed the old way of dispatching events and used entt::dispatcher
- event dispatcher function implementation is an inl file, never knew of those

* Bug in EventDispatcher::removeListener, function needed an Event& + comments

* Added DirectoryWatcher

The DirectoryWatcher has a list of commands (basically callbacks) that are executed depending on the type of event in the swtich, FileCreatedEvent, FileDeletedEvent and so on, i won't implement directory cause it would be redundant for now... maybe later i'll find some edge case that will force me to do that.

I get the filename and path to the files from the watcher, more things i can do with the asset manager, like hot reloading scripts or textures and so on.
Currently it looks like the directory watcher fires extra events and must make sure i get things right with hot reloading.

Directory navigation in the Assets window, will implement drag and drop to add elements to the current scene

* Added drag and drop for files into the viewport

It just prints the path of the file, no loading involved, might need to filter the files depending on what the engine can load... for example for now it knows how to load only textures (limited to types of files SOIL can parse) and models (only gltf)

* Using spdlog instead of my own logger

I could not get it to work with format for the life of me, why should i reinvent the wheel (this specific wheel)

* Bug with log, have to be set_default before init() call

* Fix crash on scene hierarchy scroll

* Design overhaul

Added iosevka font and had a look at the cherno theme for the imgui, the file explorer now has folder navigation and file drag and drop functionality.

* Adding entities/modifying based on asset payload

Files are currently filtered by extension (jpg and png for textures and ONLY gltf for models)
- can drop model files and they will get loaded
- can add a texture and it will automatically create an entity and add the texture to it
    * if we already have an entity selected it will clear the texture, load the one we dropped and
      assign it to the entity

* Implemented a simple RenderSystem

It looks ugly but at the moment it gets the job done
Model loading and drawing upon dropping the file into the viewport
 * need  to add object properties window to support adding components from there

* ~ Instanced rendering ~

Managed to implement instanced rendering, uploading the geometry once per model and then if we want to draw the same geometry but 100 times, we will use a single draw call to draw them all.
There is a catch with glDrawElementsInstanced, it will need a transformation matrix, one per instance so the geometry could be the same but the instance matrix should be different.

This works verry well for things like grass, or trees where we have thousands of them and need to render them efficiently.

- vertex shader is update to support instance matrix
- Model class has new functions tied to those instances, for sure the instance matrices vector SHOULD NOT live in the model class, I think at first glance that we need some counter to determine how many of the same object we must draw and then setup the instancing data

800 thousand cubes with a single texture ~100fps

* Camera moves better

Using relative coordinates for MouseMovedEvent, had to inject the SDL_Window into the scene viewport for the relative mode

* Multiple fixes + context menus

- camera had Y inverted with pitch += y * mouse sensitivity but it had to be -=
- made model component aggregate, as well as name component, now we can do ModelComponent { .loaded = true, .pModel = model}
- scene hierarchy has PER ITEM and window context menus, currently only the window context menu has minor functionality by clearing the selected entity
- file explorer now has it's own event dispatcher
- dropping unsigned int for uint32_t for visibility and clarity
- scene must manage the instance map
    * WORK IN PROGRESS
- SelectEntityEvent now supports unselecting using a default ctor SelectEntity():m_entity{entt::null}

* README update

* Need a lot of refactoring

I will commit this as is and merge with main, i don't like this branch... don't see the point to it

* FIX: FPS boost due to directory iteration

I was iterating the files over and over again even though i had no updates and with many files to render it would drop my fps down to 100