imgui_stdlib added
Emit deselect entity event when we clear the hierarchy selection and when we create a new entity.

Table instead of listbox.

Must add functionality to the Entity group since we now use IdentifierComponent to hold more data about an entity.

Added imgui_stdlib for the InputText, now instead of having a char[] buffer we can pass the address of the string so we edit the source directly