Somewhat working shadows from directional light
I managed to get shadow mapping working, shadows do not look that good and i m still trying to understand the math behind it since majority of the shader code is from the learnopengl website.

I need to work a lot on the lighting system and shadows.

DirectionalLights are not saved on quit event