Directional light now casts acutal light
Need to upload to the gpu a color too since right now it is yellow and the shadows are still not like i'd like them to be but we're getting there.
Also there is no need for cpy bat file since now i edit shaders that are located already where they should be.
MOVED TO VSCODE as editor for shader files, might migrate to another based on config or something that has glsl linting