AssetManager singleton
Due to the fact that i want to be able to use icons for the UI of the editor i need the asset manager available everywhere.

I want to draw an icon when i drag and drop meshes to load, i can do that in two ways, an unordered map of string, textureId and preload them at start time where i initialise the imgui manager oooor i make the asset manager a singleton and can use it in the imgui manager however i like.

This allows for better flexibility and should have been the decision from the get go for this specific piece of functionality of the engine.

Bug fix: textures loaded from the mesh file were using path for indexing in map, and anywhere else name was used, now everything uses path.